Add abstract JobMetadataAttribute to share ResourceManager logic between JobDescriptionAttribute and JobDisplayNameAttribute
